pyproject.toml
[project]
name = "ai-interview-preparation-platform"
version = "0.1.0"
description = ""
authors = [
    {name = "Roni Shternberg"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "anthropic (>=0.40.0,<1.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "django-ai-assistant (>=0.1.2,<0.2.0)",
    "gitpython (>=3.1.44,<4.0.0)",
    "langchain-community (>=0.3.25,<0.4.0)",
    "django-webpack-loader (>=3.2.1,<4.0.0)",
    "pycmarkgfm (>=1.2.1,<2.0.0)",
    "scikit-learn (>=1.7.0,<2.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "microsandbox (>=0.1.8,<0.2.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "code2flow (>=2.5.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
